+++ packages/eslint-plugin/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Remove `fixable` option from `i18n-no-flanking-whitespace` rule.
+
 ## 22.11.0 (2025-06-04)
 
 ## 22.10.0 (2025-05-22)